Refactor actions/reducers to use promises
Also renames some camelCase.js files into a file-name.js structure to maintain consistency with the rest of the codebase.
