+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux"
-import { signUpReducer } from "./signUpReducer"
-import { requirementReducer } from "./requirementReducer"
+import signUpReducer from "./sign-up"
+import requirementReducer from "./requirements"
 
 export default combineReducers({
   signUpReducer,
