--- src/reducers/signUpReducer.js
@@ -1,31 +1,0 @@
-import { SIGNUP_IN_PROGRESS, SIGNUP_SUCCESS, SIGNUP_FAILURE } from "../actions"
-
-const initialState = {
-  inProgress: false,
-  serverError: {},
-  isServerError: false
-}
-
-export const signUpReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case SIGNUP_IN_PROGRESS:
-      return {
-        ...state,
-        inProgress: true
-      }
-    case SIGNUP_SUCCESS:
-      return {
-        ...state,
-        inProgress: false
-      }
-    case SIGNUP_FAILURE:
-      return {
-        ...state,
-        inProgress: false,
-        serverError: action.payload,
-        isServerError: true
-      }
-    default:
-      return state
-  }
-}
